refactor: requester 패키지 변경
[#24]
